requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.0.4
autopep8==1.4
colorama==0.3.9
cycler==0.10.0
Django==2.0.5
django-tastypie==0.14.1
djangorestframework==3.8.2
flake8==3.5.0
gast==0.2.0
gcloud==0.17.0
googleapis-common-protos==1.5.5
grpcio==1.16.0
gunicorn==19.9.0
h5py==2.8.0
httplib2==0.12.0
isort==4.3.4
jws==0.1.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.1
mccabe==0.6.1
numpy==1.15.3
oauth2client==3.0.0
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.1
psycopg2==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycryptodome==3.4.3
pyflakes==1.6.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyparsing==2.2.2
Pyrebase==3.0.27
python-dateutil==2.7.3
python-jwt==2.0.1
python-mimeparse==1.6.0
pytz==2018.4
PyYAML==3.13
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
